# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MC Roleplay Server API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List support tickets (admins see all, users see their own)
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketInput"
      responses:
        "201":
          description: Created ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
...
  /admin/ticket-reasons:
    get:
      operationId: listAdminTicketReasons
      tags: [tickets]
      summary: List all support ticket reasons (admin only)
      responses:
        "200":
          description: List of all ticket reasons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TicketReason"
    post:
      operationId: createTicketReason
      tags: [tickets]
      summary: Create a support ticket reason (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketReasonInput"
      responses:
        "201":
          description: Created ticket reason
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketReason"
  /admin/ticket-reasons/{id}:
    patch:
      operationId: updateTicketReason
      tags: [tickets]
      summary: Update a support ticket reason (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketReasonUpdate"
      responses:
        "200":
          description: Updated ticket reason
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketReason"
    delete:
      operationId: deleteTicketReason
      tags: [tickets]
      summary: Delete a support ticket reason (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted ticket reason
  /tickets/{id}:
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update ticket status or assigned admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketInput"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}/messages:
    get:
      operationId: listTicketMessages
      tags: [tickets]
      summary: List replies of a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of ticket replies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TicketMessage"
    post:
      operationId: sendTicketMessage
      tags: [tickets]
      summary: Send a reply/message inside a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendTicketMessageInput"
      responses:
        "201":
          description: Created reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketMessage"
...
components:
  schemas:
...
    Ticket:
      type: object
      required: [id, creatorId, reason, description, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        creatorId:
          type: integer
        reason:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        adminId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        creatorUsername:
          type: ["string", "null"]
        creatorDisplayName:
          type: ["string", "null"]
        adminUsername:
          type: ["string", "null"]
        adminDisplayName:
          type: ["string", "null"]
    CreateTicketInput:
      type: object
      required: [reason, description]
      properties:
        reason:
          type: string
        description:
          type: string
          minLength: 5
          maxLength: 2000
    TicketReason:
      type: object
      required: [id, label, isActive, order, createdAt, updatedAt]
      properties:
        id:
          type: integer
        label:
          type: string
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        order:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    TicketReasonInput:
      type: object
      required: [label]
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 500
        isActive:
          type: boolean
        order:
          type: integer
    TicketReasonUpdate:
      type: object
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 500
        isActive:
          type: boolean
        order:
          type: integer
    UpdateTicketInput:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        adminId:
          type: ["integer", "null"]
    TicketMessage:
      type: object
      required: [id, ticketId, senderId, content, createdAt, senderUsername]
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        senderId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
        senderUsername:
          type: string
        senderDisplayName:
          type: ["string", "null"]
        senderAvatarUrl:
          type: ["string", "null"]
    SendTicketMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
...