# lib/api-zod/src/generated/types/sendTicketMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MC Roleplay Server API
 * OpenAPI spec version: 0.1.0
 */

export interface SendTicketMessageInput {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content: string;
}